Добавляет загрузку изображения
