+++ js/comments.js
@@ -1,15 +1,20 @@
-import { isEscapeKey } from './util.js';
-
 const COMMENTS_LOAD_STEP = 5;
 const bigPicture = document.querySelector('.big-picture');
 const commentsList = bigPicture.querySelector('.social__comments');
 const commentTemplate = commentsList.querySelector('.social__comment');
-const closeBigPictureButton = bigPicture.querySelector('.big-picture__cancel');
 const loadCommentsButton = bigPicture.querySelector('.comments-loader');
 const сommentsCountData = bigPicture.querySelector('.social__comment-count');
 
 let displayedCommentsCount = COMMENTS_LOAD_STEP;
 let allComments = [];
+
+const replaceAllComments = (newComments = []) => {
+  allComments = newComments;
+};
+
+const updateDisplayedCommentsCount = (newCount) => {
+  displayedCommentsCount = newCount;
+};
 
 const createCommentElement = (comment) => {
   const commentElement = commentTemplate.cloneNode(true);
@@ -47,49 +52,4 @@ const onLoadCommentsButtonClick = () => {
   renderComments();
 };
 
-const closeBigPicture = () => {
-  bigPicture.classList.add('hidden');
-  document.body.classList.remove('modal-open');
-
-  displayedCommentsCount = COMMENTS_LOAD_STEP;
-  allComments = [];
-};
-
-const onEscapeKeyDown = (evt) => {
-  if (isEscapeKey(evt)) {
-    closeBigPicture();
-
-    document.removeEventListener('keydown', onEscapeKeyDown);
-    loadCommentsButton.removeEventListener('click', onLoadCommentsButtonClick);
-  }
-};
-
-const onCloseBigPictureClick = () => {
-  closeBigPicture();
-
-  document.removeEventListener('keydown', onEscapeKeyDown);
-  loadCommentsButton.removeEventListener('click', onLoadCommentsButtonClick);
-};
-
-const addEventListeners = () => {
-  document.addEventListener('keydown', onEscapeKeyDown);
-  loadCommentsButton.addEventListener('click', onLoadCommentsButtonClick);
-};
-
-const openBigPicture = (picture) => {
-  bigPicture.classList.remove('hidden');
-  bigPicture.querySelector('.big-picture__img img').src = picture.url;
-  bigPicture.querySelector('.likes-count').textContent = picture.likes;
-  bigPicture.querySelector('.comments-count').textContent = picture.comments.length;
-  bigPicture.querySelector('.social__caption').textContent = picture.description;
-  allComments = picture.comments.slice();
-
-  renderComments();
-  addEventListeners();
-
-  document.body.classList.add('modal-open');
-};
-
-closeBigPictureButton.addEventListener('click', onCloseBigPictureClick);
-
-export { openBigPicture };
+export { replaceAllComments, updateDisplayedCommentsCount, renderComments, onLoadCommentsButtonClick };
