+++ js/effects-slider.js
@@ -1,62 +1,30 @@
-const effectRadios = document.querySelectorAll('.effects__radio');
-const imagePreview = document.querySelector('.img-upload__preview img');
-const effectLevelContainer = document.querySelector('.img-upload__effect-level');
-const effectLevelSlider = document.querySelector('.effect-level__slider');
-const effectLevelValue = document.querySelector('.effect-level__value');
+import { applyEffect, resetEffect } from './image-effects.js';
 
-const effectOptions = {
+const EFFECT_OPTIONS = {
   chrome: { range: { min: 0, max: 1 }, step: 0.1, start: 1 },
   sepia: { range: { min: 0, max: 1 }, step: 0.1, start: 1 },
   marvin: { range: { min: 0, max: 100 }, step: 1, start: 100 },
   phobos: { range: { min: 0, max: 3 }, step: 0.1, start: 3 },
   heat: { range: { min: 1, max: 3 }, step: 0.1, start: 3 },
 };
+const effectRadioButtons = document.querySelectorAll('.effects__radio');
+const effectLevelContainer = document.querySelector('.img-upload__effect-level');
+const effectLevelSlider = document.querySelector('.effect-level__slider');
+const effectLevelValue = document.querySelector('.effect-level__value');
 
 noUiSlider.create(effectLevelSlider, {
-  range: {
-    min: 0,
-    max: 1
-  },
+  range: { min: 0, max: 1 },
   start: 1,
   step: 0.1,
   connect: 'lower',
   format: {
     to: (value) => value.toFixed(1),
-    from: (value) => parseFloat(value)
-  }
+    from: (value) => parseFloat(value),
+  },
 });
 
-const resetEffect = () => {
-  effectLevelContainer.style.display = 'none';
-  imagePreview.style.filter = '';
-  effectLevelValue.value = '';
-};
-
-const applyEffect = (effect, value) => {
-  switch (effect) {
-    case 'chrome':
-      imagePreview.style.filter = `grayscale(${value})`;
-      break;
-    case 'sepia':
-      imagePreview.style.filter = `sepia(${value})`;
-      break;
-    case 'marvin':
-      imagePreview.style.filter = `invert(${value}%)`;
-      break;
-    case 'phobos':
-      imagePreview.style.filter = `blur(${value}px)`;
-      break;
-    case 'heat':
-      imagePreview.style.filter = `brightness(${value})`;
-      break;
-    default:
-      resetEffect();
-      break;
-  }
-};
-
 const updateSliderOptions = (effect) => {
-  const options = effectOptions[effect];
+  const options = EFFECT_OPTIONS[effect];
   effectLevelSlider.noUiSlider.updateOptions({
     range: options.range,
     step: options.step,
@@ -75,7 +43,7 @@ const onEffectChange = () => {
   }
 };
 
-effectRadios.forEach((effect) => effect.addEventListener('change', onEffectChange));
+effectRadioButtons.forEach((effect) => effect.addEventListener('change', onEffectChange));
 
 effectLevelSlider.noUiSlider.on('update', (values, handle) => {
   const effect = document.querySelector('.effects__radio:checked').value;
@@ -83,5 +51,3 @@ effectLevelSlider.noUiSlider.on('update', (values, handle) => {
   effectLevelValue.value = value;
   applyEffect(effect, value);
 });
-
-export { resetEffect };
