+++ js/form-handler.js
@@ -2,8 +2,9 @@ import { isEscapeKey } from './util.js';
 import { initializeValidation } from './validation.js';
 import { submitForm } from './form-submit.js';
 import { resetScale } from './image-scale.js';
-import { resetEffect } from './effects-slider.js';
+import { resetEffect } from './image-effects.js';
 
+const FILE_TYPES = ['jpg', 'jpeg', 'png'];
 const uploadForm = document.querySelector('.img-upload__form');
 const imageUploadInput = document.querySelector('.img-upload__input');
 const imageEditForm = document.querySelector('.img-upload__overlay');
@@ -13,13 +14,21 @@ const effectLevel = document.querySelector('.effect-level__value');
 const hashtagsInput = document.querySelector('.text__hashtags');
 const descriptionInput = document.querySelector('.text__description');
 const effectButtons = document.querySelectorAll('.effects__radio');
+const previewImage = document.querySelector('.img-upload__preview img');
 
 let isErrorMessageOpen = false;
 
 imageUploadInput.addEventListener('change', () => {
   if (imageUploadInput.files.length > 0) {
-    imageEditForm.classList.remove('hidden');
-    document.body.classList.add('modal-open');
+    const file = imageUploadInput.files[0];
+    const fileName = file.name.toLowerCase();
+    const matches = FILE_TYPES.some((it) => fileName.endsWith(it));
+
+    if (matches) {
+      previewImage.src = URL.createObjectURL(file);
+      imageEditForm.classList.remove('hidden');
+      document.body.classList.add('modal-open');
+    }
   }
 });
