+++ js/image-scale.js
@@ -1,12 +1,11 @@
 const MIN_SCALE = 25;
 const MAX_SCALE = 100;
 const SCALE_STEP = 25;
-
 const smallerButton = document.querySelector('.scale__control--smaller');
 const biggerButton = document.querySelector('.scale__control--bigger');
 const scaleValueInput = document.querySelector('.scale__control--value');
 const imagePreview = document.querySelector('.img-upload__preview');
-let scaleValue = 100;
+let scaleValue = MAX_SCALE;
 
 const updateScale = () => {
   scaleValueInput.value = `${scaleValue}%`;
@@ -18,18 +17,15 @@ const resetScale = () => {
   updateScale();
 };
 
-smallerButton.addEventListener('click', () => {
-  if (scaleValue > MIN_SCALE) {
-    scaleValue -= SCALE_STEP;
+const changeScale = (delta) => {
+  const newScale = scaleValue + delta;
+  if (newScale >= MIN_SCALE && newScale <= MAX_SCALE) {
+    scaleValue = newScale;
     updateScale();
   }
-});
+};
 
-biggerButton.addEventListener('click', () => {
-  if (scaleValue < MAX_SCALE) {
-    scaleValue += SCALE_STEP;
-    updateScale();
-  }
-});
+smallerButton.addEventListener('click', () => changeScale(-SCALE_STEP));
+biggerButton.addEventListener('click', () => changeScale(SCALE_STEP));
 
 export { resetScale };
