+++ js/main.js
@@ -1,4 +1,6 @@
-import './pictures.js';
+import { fetchPictures } from './pictures.js';
 import './form-handler.js';
 import './image-scale.js';
 import './effects-slider.js';
+
+fetchPictures();
